+++ lib/features/workbench/pages/workbench_page/configuration/workbench_constants.dart
@@ -133,5 +133,18 @@ abstract final class WorkbenchConstants {
     0.90,
   ];
 
+  static const imageQualityRatios = [
+    0.10,
+    0.20,
+    0.30,
+    0.40,
+    0.50,
+    0.60,
+    0.70,
+    0.80,
+    0.90,
+    1.00,
+  ];
+
   static const defaultTargetSizeRatio = 0.60;
 }