+++ lib/features/workbench/pages/workbench_page/layout/task_list_card.dart
@@ -38,13 +38,10 @@ class WorkbenchTaskListCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(27, 31, 27, 0),
-      child: taskList.when(
-        loading: _buildLoading,
-        error: (error, stackTrace) => _buildError(error),
-        data: _buildList,
-      ),
+    return taskList.when(
+      loading: _buildLoading,
+      error: (error, stackTrace) => _buildError(error),
+      data: _buildList,
     );
   }
 
@@ -70,9 +67,10 @@ class WorkbenchTaskListCard extends StatelessWidget {
       itemBuilder: (context, index) {
         final task = tasks[index];
 
-        return Padding(
+        return Container(
           key: ValueKey(task.id),
           padding: EdgeInsets.only(bottom: index == tasks.length - 1 ? 0 : 13),
+          margin: EdgeInsets.fromLTRB(27, index != 0 ? 6 : 26, 27, index != tasks.length - 1 ? 0 : 48),
           child: MediaTaskListTile(
             task: task,
             selected: selectedTask?.id == task.id,